requirements.txt
# SillyMD 企业微信通道插件依赖
# 优先使用 wheels/ 目录中的离线包安装

# 核心依赖
numpy==1.24.4
onnxruntime==1.16.3
kaldi-native-fbank==1.22.3

# Whisper 依赖（备用）
torch==2.0.1
torchaudio==2.0.2
tqdm

# 网络通信
requests>=2.28.0
websockets>=11.0

# 加解密
cryptography>=41.0.0

# 配置管理
python-dotenv>=1.0.0

# 音频处理（可选，如不使用本地处理可不装）
# scipy>=1.10.0
# librosa>=0.10.0
